chore: Refactor repository and service class names
