+++ src/ProjectOrigin.Chronicler.Server/Repositories/IChroniclerRepository.cs
@@ -1,4 +1,5 @@
 using System.Threading.Tasks;
+using OpenTelemetry;
 using ProjectOrigin.Chronicler.Server.Models;
 
 namespace ProjectOrigin.WalletSystem.Server.Repositories;
