+++ src/ProjectOrigin.TestUtils/Fixtures/PostgresDatabaseFixture.cs
@@ -11,7 +11,7 @@ using ProjectOrigin.ServiceCommon.DataPersistence.Postgres;
 using Testcontainers.PostgreSql;
 using Xunit;
 
-namespace ProjectOrigin.TestUtils;
+namespace ProjectOrigin.TestUtils.Fixtures;
 
 public class PostgresDatabaseFixture<TScriptAssembly> : IAsyncLifetime
 {
